UADItem: Fixed temporary variable bug. UBAShowObjects: UBAShowObjectsSimple class created. UBACamera & UBACameraSimple: Major changes. Child classes may be not work! Core/Math: Major changes in MVector classes. MVector inherited from MMatrix now. myrdk.h: IVisualInterface renamed to UIVisualController UBADataSimulator: Created. GUI/Builder: minor changes.

+++ GUI/Builder/UClassesListFrameUnit.cpp
@@ -32,7 +32,7 @@ void TUClassesListFrame::UpdateInterface(void)
  std::vector<int> ids;
  ids.resize(numclasses);
  Storage_GetClassesList(&ids[0]);
- for(size_t i=0;i<numclasses;i++)
+ for(int i=0;i<numclasses;i++)
  {
   StringGrid->Cells[0][i+1]=IntToStr(ids[i]);
   StringGrid->Cells[1][i+1]=Storage_GetClassName(ids[i]);
